Refactor: Deduplicate transport bitrate estimation logic
Extracts the EMA logic from updateTransportBitrate into a reusable TransportBitrateEstimator utility class to avoid duplication between SRT regulators.
